requirements.txt
argh==0.26.2
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
nose==1.3.7
numpy==1.13.1
oauth2client==4.1.2
olefile==0.44
pathtools==0.1.2
Pillow==4.3.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pytesseract==0.1.7
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
slackclient==1.0.9
uritemplate==3.0.0
urllib3==1.22
watchdog==0.8.3
websocket-client==0.44.0
